Try moving content layout down to the layout type, combine with contentSize and wideSize
