+++ packages/block-editor/src/hooks/layout.scss
@@ -6,7 +6,6 @@
 
 .block-editor-hooks__layout-controls {
 	display: flex;
-	margin-bottom: $grid-unit-10;
 
 	.block-editor-hooks__layout-controls-unit {
 		display: flex;
